+++ lib/rewrites.php
@@ -28,8 +28,8 @@ function roots_add_rewrites($content) {
 }
 
 function roots_clean_urls($content) {
-  if (strpos($content, FULL_RELATIVE_PLUGIN_PATH) === 0) {
-    return str_replace(FULL_RELATIVE_PLUGIN_PATH, WP_BASE . '/plugins', $content);
+  if (strpos($content, RELATIVE_PLUGIN_PATH) > 0) {
+    return str_replace('/' . RELATIVE_PLUGIN_PATH,  '/plugins', $content);
   } else {
     return str_replace('/' . THEME_PATH, '', $content);
   }